传输页面 修复 在对方主动关闭会话后，子连接仍然尝试重连

+++ src/utils/peer/FileTransfer.js
@@ -110,7 +110,7 @@ export class FileTransfer {
         })
 
         conn.on('close', () => {
-            if(!this.closed) {
+            if(!this.closed && this.mainConn._open) {
                 console.error('subconn closed unexpectedly: ', conn)
                 if(this.fileReaderWorker) conn.fileReaderWorker.terminate()
                 this.subConns.splice(this.subConns.indexOf(conn), 1)
